refactor: 파일 패키지 정리 및 에러 클래스 리팩터링
