+++ app/src/main/java/woowacourse/shopping/data/dummy/DummyProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.remote
+package woowacourse.shopping.data.dummy
 
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.domain.repository.ProductRepository
