+++ app/src/main/java/woowacourse/shopping/presentation/ui/shopping/ShoppingError.kt
@@ -8,5 +8,4 @@ enum class ShoppingError(
     RecentProductItemsNotFound("최근 본 상품을 불러올 수 없습니다."),
     CartItemsNotFound("장바구니를 불러올 수 없습니다."),
     CartItemsNotModified("상품 수량을 변경할 수 없습니다."),
-    CartItemNotDeleted("장바구니에서 상품을 지울 수 없습니다."),
 }
